+++ CHANGELOG.md
@@ -9,6 +9,9 @@ x.x.x Release notes (yyyy-MM-dd)
 * Fixed a crash after clearing a list or set of Mixed containing links to objects. ([realm/realm-core#4774](https://github.com/realm/realm-core/issues/4774), since the beginning of v11)
 * Fixed an endless recursive loop that could cause a stack overflow when computing changes on a set of objects which contained cycles. ([#4770](https://github.com/realm/realm-core/pull/4770), since the beginning of v11)
 
+### Known Issues
+* `instanceof Realm.Dictionary` will always be `false` on React Native. ([#3836](https://github.com/realm/realm-js/issues/3836))
+
 ### Compatibility
 * MongoDB Realm Cloud.
 * Realm Studio v11.0.0.